# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deliverance Church Lugazi (DCL) Church Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      tags: [People]
      operationId: listPeople
      summary: List pastors / ministers / members (filterable)
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PersonRole'
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
    post:
      tags: [People]
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
  /people/{id}:
    get:
      tags: [People]
      operationId: getPerson
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
    patch:
      tags: [People]
      operationId: updatePerson
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
    delete:
      tags: [People]
      operationId: deletePerson
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
  /finance/transactions:
    get:
      tags: [Finance]
      operationId: listFinanceTransactions
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
        - name: kind
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FinanceKind'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FinanceTransaction'
    post:
      tags: [Finance]
      operationId: createFinanceTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceTransactionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
  /finance/transactions/{id}:
    patch:
      tags: [Finance]
      operationId: updateFinanceTransaction
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceTransactionInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
    delete:
      tags: [Finance]
      operationId: deleteFinanceTransaction
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
  /social-links:
    get:
      tags: [Social]
      operationId: listSocialLinks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SocialLink'
    post:
      tags: [Social]
      operationId: createSocialLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLinkInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialLink'
  /social-links/{id}:
    patch:
      tags: [Social]
      operationId: updateSocialLink
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLinkInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialLink'
    delete:
      tags: [Social]
      operationId: deleteSocialLink
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    PersonRole:
      type: string
      enum: [pastor, minister, member]
    FinanceKind:
      type: string
      enum: [tithe, offering, pledge, donation, expense, other_income]
...
    Person:
      type: object
      required: [id, fullName, role, isLeader, branchId, departmentId, email, phone, photoUrl, bio, joinedOn, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/PersonRole'
        isLeader:
          type: boolean
        branchId:
          type: [integer, 'null']
        departmentId:
          type: [integer, 'null']
        email:
          type: [string, 'null']
          format: email
        phone:
          type: [string, 'null']
        photoUrl:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        joinedOn:
          type: [string, 'null']
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PersonInput:
      type: object
      required: [fullName, role]
      properties:
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/PersonRole'
        isLeader:
          type: boolean
        branchId:
          type: [integer, 'null']
        departmentId:
          type: [integer, 'null']
        email:
          type: [string, 'null']
          format: email
        phone:
          type: [string, 'null']
        photoUrl:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        joinedOn:
          type: [string, 'null']
          format: date
...
    FinanceTransaction:
      type: object
      required: [id, kind, amount, currency, branchId, occurredOn, description, recordedBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        kind:
          $ref: '#/components/schemas/FinanceKind'
        amount:
          type: number
        currency:
          type: string
        branchId:
          type: [integer, 'null']
        occurredOn:
          type: string
          format: date
        description:
          type: [string, 'null']
        recordedBy:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FinanceTransactionInput:
      type: object
      required: [kind, amount, occurredOn]
      properties:
        kind:
          $ref: '#/components/schemas/FinanceKind'
        amount:
          type: number
        currency:
          type: string
        branchId:
          type: [integer, 'null']
        occurredOn:
          type: string
          format: date
        description:
          type: [string, 'null']
        recordedBy:
          type: [string, 'null']
...
    SocialLink:
      type: object
      required: [id, platform, url, handle, isActive, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        platform:
          type: string
        url:
          type: string
        handle:
          type: [string, 'null']
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SocialLinkInput:
      type: object
      required: [platform, url]
      properties:
        platform:
          type: string
        url:
          type: string
        handle:
          type: [string, 'null']
        isActive:
          type: boolean
        sortOrder:
          type: integer
...